+++ public/js/ws.js
@@ -8,6 +8,7 @@ import { refreshTabMeta } from "./tabs.js";
 import { setCustomCommands, refreshOpenPalette } from "./palette.js";
 import { setLayouts } from "./layouts.js";
 import { setGitStatus, setGitDiff } from "./git.js";
+import { setMdContent } from "./md-preview.js";
 import { setShortcutOverrides } from "./shortcuts.js";
 import { refreshOpenSettings } from "./settings.js";
 import { setAgents } from "./agents.js";
@@ -141,6 +142,9 @@ export function connect() {
         break;
       case "gitDiff":
         setGitDiff(msg);
+        break;
+      case "fileContent":
+        setMdContent(msg);
         break;
       case "settings":
         setShortcutOverrides(msg.settings && msg.settings.shortcuts);